COMP: warning fixes
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=10608
git-svn-id: http://svn.slicer.org/Slicer4/trunk@10608 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/GUI/Testing/EventBrokerTest.cxx
@@ -47,7 +47,7 @@ int main(int argc, char * argv[])
   callback->SetCallback( Callback ); 
   callback->SetClientData( reinterpret_cast<void *> (viewer) );
 
-  int i, numberOfObservations = 1;
+  size_t i, numberOfObservations = 1;
   for (i = 0; i < numberOfObservations; i++)
     {
     broker->AddObservation( ellip, vtkCommand::ModifiedEvent, viewer, callback);
